+++ src/components/media/Video/utils/logWarning.js
@@ -1,0 +1,6 @@
+export const logWarning = videoSource => {
+  // eslint-disable-next-line no-console
+  console.error(
+    `Warning:  \`Video.props.videoSource\` must be a valid URL or valid HTML. Received: ${videoSource}`
+  );
+};